Restructure app with API versioning
