--- app/controllers/comments_controller.rb
@@ -1,41 +1,0 @@
-class CommentsController < ApplicationController
-  before_action :set_post
-  before_action :set_post_comment, only: [:show, :update, :destroy]
-
-  def index
-    json_response(@post.comments)
-  end
-
-  def show
-    json_response(@comment)
-  end
-
-  def create
-    @post.comments.create!(comment_params)
-    json_response(@post, :created)
-  end
-
-  def update
-    @comment.update(comment_params)
-    head :no_content
-  end
-
-  def destroy
-    @comment.destroy
-    head :no_content
-  end
-
-  private
-
-  def comment_params
-    params.permit(:content, :created_by)
-  end
-
-  def set_post
-    @post = Post.find(params[:post_id])
-  end
-
-  def set_post_comment
-    @comment = @post.comments.find_by!(id: params[:id]) if @post
-  end
-end
