+++ config/routes.rb
@@ -1,10 +1,14 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
-  resources :posts do
-    resources :comments
-  end
+  namespace :api do
+    namespace :v1 do
+      resources :posts do
+        resources :comments
+      end
 
-  post 'auth/login', to: 'authentication#authenticate'
-  post 'signup', to: 'users#create'
+      post 'auth/login', to: 'authentication#authenticate'
+      post 'signup', to: 'users#create'
+    end
+  end
 end
