+++ spec/auth/api/v1/authenticate_user_spec.rb
