+++ spec/requests/api/v1/users_spec.rb
@@ -4,28 +4,32 @@ RSpec.describe 'Users API', type: :request do
   let(:user) { build(:user) }
   let(:headers) { valid_headers.except('Authorization') }
   let(:valid_attributes) do
-    attributes_for(:user, password_confirmation: user.password)
+    attributes_for(
+      :user,
+      password: user.password_digest,
+      password_confirmation: user.password_digest
+    )
   end
 
   describe 'POST /signup' do
     context 'when valid request' do
-      before { post '/signup', params: valid_attributes.to_json, headers: headers }
+      before { post api_v1_signup_path, params: valid_attributes.to_json, headers: headers }
 
-      # it 'creates a new user' do
-      #   expect(response).to have_http_status(201)
-      # end
+      it 'creates a new user' do
+        expect(response).to have_http_status(201)
+      end
 
-      # it 'returns success message' do
-      #   expect(json['message']).to match(/Account created successfully/)
-      # end
+      it 'returns success message' do
+        expect(json['message']).to match(/Account created successfully/)
+      end
 
-      # it 'returns an authentication token' do
-      #   expect(json['auth_token']).not_to be_nil
-      # end
+      it 'returns an authentication token' do
+        expect(json['auth_token']).not_to be_nil
+      end
     end
 
     context 'when invalid request' do
-      before { post '/signup', params: {}, headers: headers }
+      before { post api_v1_signup_path, params: {}, headers: headers }
 
       it 'does not create a new user' do
         expect(response).to have_http_status(422)
